pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "open-clip-torch (>=2.31.0,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langfuse (>=2.59.7,<3.0.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langchain-google-genai (>=2.1.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
